shuffling files around
